requirements.txt
streamlit
sentence-transformers
scikit-learn
pandas
nltk
PyPDF2
python-dotenv
openai